feat(elaborate): Add checks for exclusive/exhaustive matches
Checks whether the rows of a table function are exclusive and
exhaustive, based on the input signature.
A separate module `Pattern` constructs sets of variants and operates
on lists of such sets (tuple patterns).
Exclusiveness checks are done by checking the intersection for each
position.
Exhaustiveness checks are done by subtracting each tuple-covered
region from the total set iteratively.
The final tablerow is split out before the checks if it is a "closer"
(all-wildcard pattern) and special-cases it for exhaustiveness.